% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_clust.R
\name{hier_clust}
\alias{hier_clust}
\title{Agglomerative Hierarchical Clustering Method`}
\usage{
hier_clust(data, method = "euclidean")
}
\arguments{
\item{data}{A data frame}

\item{method}{A string that states which distance method the user wants to use}
}
\value{
A list with clustering path
}
\description{
Agglomerative Hierarchical Clustering Method`
}
