% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_means.R
\name{k_means_plus_plus}
\alias{k_means_plus_plus}
\title{Run K-means ++ initialization method to determine initial centroid points}
\usage{
k_means_plus_plus(dat, k)
}
\arguments{
\item{dat}{A data set}

\item{k}{Number of clusters}
}
\value{
Three points in the data set that will start as the inital centroids for each cluster
}
\description{
Run K-means ++ initialization method to determine initial centroid points
}
