% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_means.R
\name{plot_clusterings}
\alias{plot_clusterings}
\title{Plots the results of k_means() clustering in the first two PC dimensions}
\usage{
plot_clusterings(cluster_assignments)
}
\arguments{
\item{cluster_assignments}{Clustering from k_means()}
}
\value{
A plot
}
\description{
Plots the results of k_means() clustering in the first two PC dimensions
}
