% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_means.R
\name{total_sum_squares}
\alias{total_sum_squares}
\title{Retrieves Total Sum of Squares}
\usage{
total_sum_squares(center, dat)
}
\arguments{
\item{center}{The center point of the data that all Euclidean distances will be calculated on}

\item{dat}{A data set}
}
\description{
Retrieves Total Sum of Squares
}
