% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_means.R
\name{k_means}
\alias{k_means}
\title{Implements K means by hands}
\usage{
k_means(dat, k, pca = FALSE)
}
\arguments{
\item{dat}{A data set}

\item{k}{The number of desired clusters}

\item{pca}{Boolean value indicating whether or not to run Principal Component Analysis prior to K means}
}
\description{
Implements K means by hands
}
